Added safety check to Update-PoshGitHub
- Since we're doing a git reset --hard HEAD make sure we're good
- git pull was 'async' before - wait on the output before reloading
the module

+++ Posh-Github.psm1
@@ -402,14 +402,24 @@ function Get-GitHubPullRequests
 
 function Update-PoshGitHub
 {
+  #$null if we can't find module (not sure how that happens, but just in case!)
   $installedPath = Get-Module Posh-GitHub |
     Select -ExpandProperty Path |
     Split-Path
 
-  Write-Host "Found Post-GitHub module at $installedPath"
   Push-Location $installedPath
+
+  #DANGER - be safe and abort as git reset --hard could do damage in wrong spot
+  if (($installedPath -eq $null) -or ((Get-Location) -ne $installedPath))
+  {
+    Pop-Location
+    throw "Could not find Posh-GitHub module / reset path - Update aborted!"
+    return
+  }
+
+  Write-Host "Found Post-GitHub module at $installedPath"
   git reset --hard HEAD | Out-Null
-  git pull
+  git pull | Write-Host
   Pop-Location
   Remove-Module Posh-GitHub
   Import-Module (Join-Path $installedPath 'Posh-GitHub.psm1')
